Nested arg, level, and param within Options
